+++ src/Classification/Classifier/Cms/MoodleClassifier.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Startwind\WebInsights\Classification\Classifier\Cms;
+
+use Startwind\WebInsights\Classification\Classifier\Classifier;
+use Startwind\WebInsights\Classification\Classifier\Http\Http\ProgrammingLanguageClassifier;
+use Startwind\WebInsights\Response\HttpResponse;
+
+class MoodleClassifier implements Classifier
+{
+    private const TAG = CmsClassifier::CLASSIFIER_PREFIX . 'moodle';
+
+    public function classify(HttpResponse $httpResponse, array $existingTags): array
+    {
+        if ($httpResponse->getHtmlDocument()->containsAny([
+                'yui-moodlesimple-min.css',
+                'moodle-core-'
+            ]
+        )) {
+            return [self::TAG, ProgrammingLanguageClassifier::TAG_PHP];
+        }
+
+        return [];
+    }
+
+}